create car

+++ homework_02/engine.py
@@ -8,4 +8,4 @@ create dataclass `Engine`
 @dataclass()
 class Engine:
     volume: float
-    pistons: float
+    pistons: int
